splitup topology JSON stats and factory
